+++ app/components/impulse/autocomplete/option_component.rb
@@ -1,11 +1,12 @@
 module Impulse
   module Autocomplete
     class OptionComponent < ApplicationComponent
-      attr_reader :value, :text, :system_args
+      attr_reader :value, :text, :description, :system_args
 
-      def initialize(value:, text:, **system_args)
+      def initialize(value:, text:, description: nil, **system_args)
         @value = value
         @text = text
+        @description = description
         @text_id = self.class.generate_id
         @system_args = system_args
         @system_args[:id] = system_args.fetch(:id, self.class.generate_id)